Handle reconnect for STAN clients
With the NATS Streaming server running in memory mode, it loses
all information about previously connected clients and clusters when
it is restarted.  This behaviour requires connected clients to
re-initialize the STAN connections when a restart is detected. This
commit adds reconnect handlers on the NATS connections that
re-initialize the STAN connections.

Signed-off-by: Vincent Smith <vosmith08@gmail.com>