Forward HTTP headers when using a callback
HTTP headers were not being forwarded correctly as identified by
Philippe Soares on Slack while using callbacks.

This change forwards the headers of the function's response to the
endpoint specified in X-Callback-Url. I tested this with netcat
running on my local machine and a new version of the queue worker
running on Docker Swarm.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>