Forward X-Call-Id from producer
This means we can trace an asynchronous call from the moment it
is queued until the callback URL receives a confirmation or error

https://github.com/openfaas/faas/issues/813

Pair-programmed with Ivana. Tested with new gateway version that
passes through new header and also with older gateway that does
not. Worked as expected on Docker Swarm. Unit tests pass.

Closes https://github.com/openfaas/faas/issues/813

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>