Add gateway_invoke feature
This change introduces the gateway_invoke env-var which is used
to have the queue-worker invoke functions via the gateway. When
set to false the functions are invoked directly over HTTP.

Tested on Swarm with the flag on/off. Unit tests added for
parsing new configuration option.

Options documented in README.md

Fixes: #32

Signed-off-by: Alex Ellis <alexellis2@gmail.com>