Add missing error handling
Due to an issue described in #73 by @bmcstdio the error handling
has been improved to prevent nil pointer dereference errors in
several scenarios.

Improves error messages for queue-worker output.

Tested with and without gateway_invoke by changing the gateway
port to an invalid one and the gateway address to an invalid one.

Closes: #73

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>